+++ src/main/java/open/commons/ssh/IConnectionInfo.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2020 Park Jun-Hong_(parkjunhong77/google/com)
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ *
+ * This file is generated under this project, "open-commons-ssh".
+ *
+ * Date  : 2020. 11. 26. 오후 5:59:38
+ *
+ * Author: Park_Jun_Hong_(fafanmama_at_naver_com)
+ * 
+ */
+
+package open.commons.ssh;
+
+/**
+ * 
+ * @since 2020. 11. 26.
+ * @version _._._
+ * @author Park_Jun_Hong_(fafanmama_at_naver_com)
+ */
+public interface IConnectionInfo {
+
+    public String getHost();
+
+    public String getPassword();
+
+    public int getPort();
+
+    public String getUsername();
+}